feat: Migrate NGINX Gateway Fabric documentation into repository (#81)
With the newly open source documentation repository, we can now
consolidate all NGINX documentation into one place. This has a lots of
benefits which include but are not limited to more effective content
re-use, pipeline engineering and style enforcement.

This commit adds all of the content for NGINX Gateway Fabric, and
updates all relevant folder and file names, including link references.
Once this content has been successfully deployed, a PR to remove all
documentation traces and tooling will be opened in the Gateway Fabric
repository, decoupling the documentation fully.

It also updates the frontmatter of every single page in the 
documentation set. They are now all consistent in their fields and 
formatting, and a handful of IA or heading changes have 
been made as low-hanging fruit to bring the documentation 
set closer to full consistency.