feat: Re-order checklist, add conventional commits, re-phrase items
This commit re-orders and re-phrases certain items in the checklist to
group them logically and make the formatting more consistent. It also
adds a footnote to the message about potentially sensitive data.

The note about potentially sensitive data now has unique phrasing for
the start of the sentence: we could potentially bold it if it needed
even more superliminal attention.