feat: Add shortcode for the latest version of NGINX Gateway Fabric (#203)
This commit adds a new shortcode to the layouts folder to store a string
representing the latest version of NGINX Gateway Fabric. This means that
instead of updating each file individually during a new release, only
the shortcode file needs to be modified.

It clarifies the versioning for documentation, which makes it easier to
review - and also reduces the amount of drift when updating, in the case
that a reference was overlooked. There may be old references that need
to be updated to use this shortcode.