feat: Update NGF Get started guide and installation documents (#224)
This commit adds explicit guidance at the top of the Get started guide,
directing the reader towards the Installation section for a production
deployment. It also removes mention of the experimental channel, which
is an unnecessary use case for trying NGF.

It also creates uniformity in the Helm and Manifest deployment
documentation, by matching how sections are formatted, and creating an
include file for identical instructions on upgrading the API resources.