feat: Update closed documentation process (#323)
This pull request updates the closed documentation contribution 
process by fixing the remote git command, and grouping the 
relevant commands a user would likely do in quick succession.

It also rephrases it on the premise that the user likely already has 
the main repository cloned, so gives and frames adding the 
remote as an earlier step in the process overall.