feat: Update closed contribution documentation with new branch (#330)
Thanks to some follow up work by @eepifanova, we can now target the main
branch for closed content work. This commit updates the documentation to
reference the main branch, and also adds specificity of origin and
remote when referencing it for clarity.

It also adds missing punctuation, and adjusts spacing in an early
paragraph using a line break.